requirements_no_audio.txt
humanize
git+https://github.com/Rapptz/discord-ext-menus.git
discord.py
pygit2
asyncpraw
aiohttp
click
spotipy
pymongo
dnspython
rule34
saucenao_api